refactor (InputView) : step3 피드백 반영 - View의 요구사항은 모두 View에서 해결하도록 수정
